refactor proxy config package
